requirements/dev.txt
-r base.txt

pytest-django==4.2.0
faker==8.1.0
factory-boy==3.2.0